requirements.txt
streamlit
py-trello
langchain
beautifulsoup4
